Moved Transports/Utilities -> Utils, moved RiptideLogger to Utils folder
