+++ RiptideNetworking/RiptideNetworking/Client.cs
@@ -4,6 +4,7 @@
 // For additional information please see the included LICENSE.md file or view it on GitHub: https://github.com/tom-weiland/RiptideNetworking/blob/main/LICENSE.md
 
 using RiptideNetworking.Transports;
+using RiptideNetworking.Utils;
 using System;
 using System.Collections.Generic;
 using System.Linq;
