+++ RiptideNetworking/RiptideNetworking/Transports/RudpTransport/RudpConnection.cs
@@ -3,6 +3,7 @@
 // Copyright (c) 2021 Tom Weiland
 // For additional information please see the included LICENSE.md file or view it on GitHub: https://github.com/tom-weiland/RiptideNetworking/blob/main/LICENSE.md
 
+using RiptideNetworking.Utils;
 using System;
 using System.Net;
