+++ RiptideNetworking/RiptideNetworking/Transports/RudpTransport/RudpPeer.cs
@@ -3,7 +3,7 @@
 // Copyright (c) 2021 Tom Weiland
 // For additional information please see the included LICENSE.md file or view it on GitHub: https://github.com/tom-weiland/RiptideNetworking/blob/main/LICENSE.md
 
-using RiptideNetworking.Transports.Utils;
+using RiptideNetworking.Utils;
 using System;
 using System.Collections.Generic;
 using System.Net;
